Added an acknowledgement system.
Added an optional acknowledgement system. Set ack.retryTimeout to a positive value to enable acknowledgements. Whether using acknowledgements or not, the code is backwards compatible with nodes running v0.3.x. However, if a node is using v0.3.x, it will not be aware of the acknowledgements and will respond to only "Marco" events.